requirements.txt
# XYZ Chatbot — Conversational RAG Pipeline
# Core ML / NLP
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4

# Dimensionality reduction & visualization
umap-learn>=0.5.6
scikit-learn>=1.4.0
plotly>=5.20.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
scipy>=1.12.0

# PNG export for Plotly figures (optional but recommended)
kaleido>=0.2.1

# LLM generation backends
openai>=1.0.0
transformers>=4.40.0
torch>=2.2.0

# Utilities
python-dotenv>=1.0.0
